backend_v2/pyproject.toml
[tool.poetry]
name = "backend-v2"
version = "0.1.0"
description = ""
authors = ["samiusss <sasab21@ulaval.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework-simplejwt = "^5.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
